+++ Sources/BraintreePayPal/BTPayPalVaultRequest.swift
@@ -6,18 +6,8 @@ import BraintreeCore
 
 ///  Options for the PayPal Vault flow.
 @objcMembers public class BTPayPalVaultRequest: BTPayPalVaultBaseRequest {
-
-    // MARK: - Public Properties
-
-    /// Optional: User email to initiate a quicker authentication flow in cases where the user has a PayPal Account with the same email.
-    public var userAuthenticationEmail: String?
-
+    
     // MARK: - Internal Properties
-
-    /// Optional: Used to determine if the customer will use the PayPal app switch flow.
-    /// Defaults to `false`.
-    /// - Warning: This property is currently in beta and may change or be removed in future releases.
-    var enablePayPalAppSwitch: Bool = false
     
     /// Optional: Recurring billing plan type, or charge pattern.
     var recurringBillingPlanType: BTPayPalRecurringBillingPlanType?
@@ -39,8 +29,11 @@ import BraintreeCore
         enablePayPalAppSwitch: Bool,
         offerCredit: Bool = false
     ) {
-        self.init(offerCredit: offerCredit, userAuthenticationEmail: userAuthenticationEmail)
-        self.enablePayPalAppSwitch = enablePayPalAppSwitch
+        self.init(
+            offerCredit: offerCredit,
+            userAuthenticationEmail: userAuthenticationEmail
+        )
+        super.enablePayPalAppSwitch = enablePayPalAppSwitch
     }
 
     /// Initializes a PayPal Vault request
@@ -57,8 +50,10 @@ import BraintreeCore
     ) {
         self.recurringBillingDetails = recurringBillingDetails
         self.recurringBillingPlanType = recurringBillingPlanType
-        self.userAuthenticationEmail = userAuthenticationEmail
-        super.init(offerCredit: offerCredit)
+        super.init(
+            offerCredit: offerCredit,
+            userAuthenticationEmail: userAuthenticationEmail
+        )
     }
 
     public override func parameters(
@@ -66,22 +61,7 @@ import BraintreeCore
         universalLink: URL? = nil,
         isPayPalAppInstalled: Bool = false
     ) -> [String: Any] {
-        var baseParameters = super.parameters(with: configuration)
-
-        if let userAuthenticationEmail, !userAuthenticationEmail.isEmpty {
-            baseParameters["payer_email"] = userAuthenticationEmail
-        }
-
-        if let universalLink, enablePayPalAppSwitch, isPayPalAppInstalled {
-            let appSwitchParameters: [String: Any] = [
-                "launch_paypal_app": enablePayPalAppSwitch,
-                "os_version": UIDevice.current.systemVersion,
-                "os_type": UIDevice.current.systemName,
-                "merchant_app_return_url": universalLink.absoluteString
-            ]
-            
-            return baseParameters.merging(appSwitchParameters) { $1 }
-        }
+        var baseParameters = super.parameters(with: configuration, universalLink: universalLink, isPayPalAppInstalled: isPayPalAppInstalled)
         
         if let recurringBillingPlanType {
             baseParameters["plan_type"] = recurringBillingPlanType.rawValue